cloud-functions/movement-alert/requirements.txt
pydantic>=2.9.2
pydantic-settings>=2.6.1
functions-framework>=3.8.1
viam-sdk==0.33.0
